Extract statements_mutations.go
